refactor and fixed unit tests
